% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMarkov.R
\name{aggregateVD}
\alias{aggregateVD}
\title{Specific client aggregation}
\usage{
aggregateVD(string_vector)
}
\arguments{
\item{string_vector}{URLs to clean}
}
\value{
The string_vector with substitutions made
}
\description{
Specific client aggregation
}

