% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMarkov.R
\name{cleanURL}
\alias{cleanURL}
\title{Replace a string with substitutions}
\usage{
cleanURL(string_vector, findme, replace = NULL, fixed = TRUE)
}
\arguments{
\item{string_vector}{vector of (URL) strings}

\item{findme}{Regex or string to find to replace}

\item{replace}{What to replace with. If NULL, uses findme string}

\item{fixed}{IF FALSE, findme is regex, if not a fixed match}
}
\value{
string_vector with replacements if required
}
\description{
Replace a string with substitutions
}

