% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMarkov.R
\name{predictMarkov}
\alias{predictMarkov}
\title{Take 3 pageviews, output Markov model prediction}
\usage{
predictMarkov(pageview_names)
}
\arguments{
\item{pageview_names}{A character vector of pageview names}
}
\value{
The prediction
}
\description{
Take 3 pageviews, output Markov model prediction
}

